Allow for exiting or killing interactive sessions
Most of the interactive sessions we are starting will exit gracefully
when asked to however byebug does not. The cleanest way to exit a byebug
sessions seems to be killing the pid however this leaves the nib
container that started it lying around so we have to `rm` that as well.