Viva la Alpine image!
Switches back to an Alpine based image for dev/ci. The result is from an
800+mb image down to 74mb and the image builds in ~1min.

Some notes:

* The currently available version of Ruby on APK is 2.3.3. This is an
upgrade from what we were previously using (2.2.x). The reason we were
using 2.2 was we wanted the gem to be backward compatible to that
version.
* Previously integration specs were relying on `nib` to be available on
the PATH. This is not the case in this image (not sure why) so I've
edited the integration specs to point to `/usr/src/app/bin/nib` via a
constant (`NIB_BIN`).