Add CircleCI configuration
At this time Codeship does not allow for exposing build details
publicly event though this is a public/open-source project.
Additionally we are able to take advantage of our `v2`
`docker-compose.yml` w/o adding additional config to define service(s).

NOTE: I've had to disable CodeClimate test coverage reporting because
the gem is expecting to access "committed at" from `git` which is not
present. The gem has an override to support his on Codeship but not for
CircleCI v2.0. :(