Upgrade Docker image
When this Dockerfile was originally defined there was not a pre-built
compose image. Having that available dramatically simplifies the
Dockerfile here. The need for upgrade comes from the fact that the
previous version of Alpine was bringing along a very old version or Ruby
(2.3) which is not longer supported and some gems were failing to
install.