Upgrade dependencies
The primary intent is to allow support for Ruby 2.7.5 and greater
(including 3.1). The only clearly incompatible code in the library is
related to Psych and the method that handles that has been updated to
support both legacy versions and the most recent (4.0).

The rest of the changes are primarily related to upgrading RuboCop and
dealing with the fall-out from that.